MAJOR LAST CHANGE
Made all app organization based for users. Users can select organization and see data relevant to that organization.

+++ cisoDashboard/src/main/java/com/tfg/cisoDashboard/service/EventService.java
@@ -1,8 +1,10 @@
 package com.tfg.cisoDashboard.service;
 
 import com.tfg.cisoDashboard.model.Event;
+import com.tfg.cisoDashboard.model.Organization;
 import com.tfg.cisoDashboard.model.User;
 import com.tfg.cisoDashboard.repository.EventRepository;
+import com.tfg.cisoDashboard.repository.OrganizationRepository;
 import com.tfg.cisoDashboard.security.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,29 +19,29 @@ public class EventService {
     private UserService userService;
     @Autowired
     private NotificationService notificationService;
+    @Autowired
+    private OrganizationRepository organizationRepository;
 
-    private Long currentUserId(){
-        Long currentUserId = SecurityUtils.getCurrentUserIdFromContext();
-        if (currentUserId == null) {
-            return null;
-        }
-        return currentUserId;
+    public List<Event> getUserEventsByOrganization(Long organizationId) {
+        return eventRepository.findByOrganizationId(organizationId);
     }
 
-    public List<Event> getUserEvents() {
-        Long userId =  currentUserId();
-        return eventRepository.findByUserId(userId);
-    }
-
-    public Event createEvent(Event event) {
+    public Event createEvent(Event event, Long organizationId) {
         User currentUser = userService.getCurrentUser();
+        Organization organization = organizationRepository.findById(organizationId)
+                .orElseThrow(() -> new RuntimeException("Organization not found"));
         event.setUser(currentUser);
-        notificationService.createNotification("New event created: " + event.getTitle(), "event");
+        event.setOrganization(organization);
+        notificationService.createNotification("New event created: " + event.getTitle(), "event", organizationId);
         return eventRepository.save(event);
     }
 
-    public void deleteEvent(Long eventId) {
+    public void deleteEvent(Long eventId, Long organizationId) {
+        Event event = eventRepository.findById(eventId)
+                .orElseThrow(() -> new RuntimeException("Event not found"));
+        if (!event.getOrganization().getId().equals(organizationId)) {
+            throw new RuntimeException("Event does not belong to the specified organization");
+        }
         eventRepository.deleteById(eventId);
     }
 }
-
